+++ samples/apps/logging/logging.cpp
@@ -1204,5 +1204,10 @@ namespace ccfapp
   {
     return make_shared<loggingapp::Logger>(nwt, context);
   }
+
+  std::vector<js::FFIPlugin> get_js_plugins()
+  {
+    return {};
+  }
   // SNIPPET_END: rpc_handler
 }
