+++ samples/apps/nobuiltins/nobuiltins.cpp
@@ -349,4 +349,9 @@ namespace ccfapp
   {
     return std::make_shared<nobuiltins::NoBuiltinsFrontend>(nwt, context);
   }
+
+  std::vector<js::FFIPlugin> get_js_plugins()
+  {
+    return {};
+  }
 }
