+++ src/apps/js_generic/js_generic_base.h
@@ -1,14 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #include "ccf/app_interface.h"
-#include "ccf/js_openenclave_plugin.h"
+
+#include <memory>
 
 namespace ccfapp
 {
-
-  std::vector<js::FFIPlugin> get_js_plugins()
-  {
-    return {js::openenclave_plugin};
-  }
-
-} // namespace ccfapp
+  std::shared_ptr<ccf::RpcFrontend> get_rpc_handler_impl(
+    ccf::NetworkTables& network, ccfapp::AbstractNodeContext& context);
+}
